Revert "ggenie.h: GG_Code.restore is now unsigned short instead of unsigned int."
This reverts commit e50b77597297f91ae10c531f68efbff777d6c8dd.

It turns out the full unsigned int is needed because 0xFFFFFFFF indicates that
the code is invalid and hasn't been set.

+++ src/gens/util/file/ggenie.h
@@ -19,7 +19,7 @@ struct GG_Code
 	char code[16];
 	char name[240];
 	unsigned int active;
-	unsigned short restore;
+	unsigned int restore;
 	unsigned int addr;
 	unsigned short data;
 };
